Fix zipping together more than 65535 files (https://github.com/nmoinvaz/minizip/issues/373)

+++ SSZipArchive/minizip/zip.c
@@ -19,6 +19,7 @@
 #include <stdlib.h>
 #include <stdint.h>
 #include <string.h>
+#include <limits.h>
 #include <errno.h>
 #include <assert.h>
 
@@ -1888,7 +1889,7 @@ extern int ZEXPORT zipClose2_64(zipFile file, const char *global_comment, uint16
     pos = centraldir_pos_inzip - zi->add_position_when_writting_offset;
 
     /* Write the ZIP64 central directory header */
-    if (pos >= UINT32_MAX || zi->number_entry > UINT32_MAX)
+    if (pos >= UINT32_MAX || zi->number_entry > UINT16_MAX)
     {
         uint64_t zip64_eocd_pos_inzip = ZTELL64(zi->z_filefunc, zi->filestream);
         uint32_t zip64_datasize = 44;
